[jnimarshalmethod-gen] Target .NET 7.
WIP WIP WIP WIP WIP

The theory so far:

`jnimarshalmethod-gen` (176240d2) worked by using
[`LambdaExpression.CompileToMethod(ILGenerator)`][0], which worked
because `JniRuntime.JniMarshalMemberBuilder` would produce a
`LambdaExpression`, and then we could use `System.Reflection` to
obtain an `ILGenerator`, and just serialize the `LambdaExpression`
into the `ILGenerator`.  This allowed us to write new assemblies
containing marshal methods that could invoke custom marshalers.

The problem is that `LambdaExpression.CompileToMethod()` never made
it into .NET Core, and is still not present as of .NET 7.

However, all of the infrastructure to write a `LambdaExpression` to
an `ILGenerator` exists, and *still* exists, within the
[dotnet/runtime repo][1].

*Copy* the dotnet/runtime System.Linq.Expressions source code into
`src-ThirdParty/dotnet-runtime`, and built it as the new assembly
`Java.Interop.Tools.ExpressionCompiler.dll`.  (This part was
surprisingly easy.)

`Java.Interop.Tools.ExpressionCompiler` is built with an
[extern alias][2] to the BCL `System.Linq.Expressions.dll`.

Add TODO-METHOD-NAME to `Java.Interop.Tools.ExpressionCompiler` which:

 1. Converts a BCL System.Linq.Expressions tree into a
    `Java.Interop.Tools.ExpressionCompiler` tree.

 2. Uses the `LambdaCompiler` infrastructure to write to an
    `ILGenerator`

 3. Update `jnimarshalmethod-gen.exe` so that it targets .NET 7, and
   uses `Java.Interop.Tools.ExpressionCompiler` instead of
   `LambdaExpression.CompileToMethod()`.

This *should* allow `jnimarshalmethod-gen` to work in .NET 7.

Assumptions:

  * (1) is possible.  (Maybe it isn't?)

  * That this is remotely sane?  The dotnet/runtime repo has this
    statement:

    > We are not accepting feature contributions to
    > System.Linq.Expressions. The library is effectively archived.

We thus have one of four possibilities:

 1. The limitations on `LambdaExpression` are "fine", and we can
    continue with this hair-brained idea.

 2. Custom marshaling is a reasonable idea, but we shouldn't use
    `LambdaExpression` and need to come up with "something else".

 3. Customer marshaling is a terrible/non-viable idea, and support
    for it should be removed from `jnimarshalmethod-gen.exe`.

 4. `jnimarshalmethod-gen.exe` as a concept should be abandoned.

[0]: https://docs.microsoft.com/en-us/dotnet/api/system.linq.expressions.lambdaexpression.compiletomethod?view=netframework-4.8
[1]: https://github.com/dotnet/runtime/tree/0d96247e828d153e738d0a2067ecbe330d7d58ab/src/libraries/System.Linq.Expressions
[2]: https://docs.microsoft.com/en-us/dotnet/csharp/language-reference/keywords/extern-alias